Reduce width of execute panel
This was the missing link for why the right side wanted to be so large.
The panel itself needed to be naturally sized instead of hard-coded. The
job name text box is also now naturally sized and adjusted based on
number of columns instead of pixels. It was made shorter as well.